Reworked unit tests
